fix(doctor,status): 同一个事实不再说两句话;三处用户读不懂的话按真机输出改掉
把用户天天看的那几个面真的读了一遍(bx status / bx doctor 的完整输出、七扇菜单
窗口的离屏快照),抓到的:

一、bx status 与 bx doctor 对同一个事实说两句不一样的话:
    status →  Notice  macOS VPN service active: Tailscale
    doctor →  [WARN]  macOS VPN service connected: 8B24B74E-… "Tailscale"  [VPN:…]
根因是 internal/supervisor/network_guard_darwin.go 与 internal/platformcheck/darwin.go
里四个函数逐字重复,同日修了前者而后者没跟上 —— 而两边的测试都绿,因为两边各测
各的那一份。三个纯解析判据下沉成叶子包 internal/macnetprobe,两处改薄壳;
darwinAnyProcessDetected 要 exec、不是解析,留在原地。

守卫 TestTheseJudgementsExistOnlyHere 断言本包之外没有第二份同样的正则(薄壳可以
很多,自己解析的不许有第二个),并带一条「本包自己那几个正则要真的在」的下限。
它当场抓到我漏删的一处残留正则,以及「新包还没 git add,git grep 看不见」这种守卫
自己失明的情形。

二、rule dead rules 打的是「14 days of cumulative uptime, short of 14 days」——
roundDays 用 %.0f 四舍五入,13.6 天被说成 14,而门槛也是 14,于是一句自相矛盾的
话。改成向下取整:在一道「还不能下结论」的门上,少说自己的进度是安全方向。

三、三处 hint 写着「answered by tunnel_claims」,而那条 check 在非 root 的报告里
根本不出现 —— 一句指向用户看不见的东西的提示,与指向不存在的命令是同一类。

四、非 root 跑 bx doctor(它刻意允许非 root)会在一份全绿的报告末尾打出
Diagnostics archive failed: mkdir …: permission denied,用户读到的是「doctor 失败
了」。现在如实说 (diagnostics archive skipped: it needs sudo bx doctor)——
「跑不了」与「跑了没过」必须分开。

二~四用新编的二进制在真机上逐条看过输出;二那条端到端看不见,因为规则体检是
Guardian 算的而这台机器的 Guardian 还跑着旧代码,由单测覆盖。

verify.sh 17 步全过。

+++ internal/rulereview/dead.go
@@ -54,11 +54,19 @@ func deadGate(in Input) (bool, string) {
 }
 
 // roundDays 把时长说成人话。报告是给人读的,`336h0m0s` 不是。
+//
+// **向下取整,不四舍五入。** 上一版用 `%.0f`,于是 13.6 天被说成 "14 days",
+// 而门槛也是 14 天 —— 真机上打出来的是
+// 「14 days of cumulative uptime, short of 14 days」,一句自相矛盾的话
+// (2026-09-18,项目所有者的 Mac)。用户读到它只会觉得这东西坏了。
+//
+// 在一道「还不能下结论」的门上,**少说自己的进度是安全方向**:说 13 天而实际
+// 13.9 天,只是让人以为还要多等一点;说 14 天而门槛是 14 天,是自相矛盾。
 func roundDays(d time.Duration) string {
 	if d < 24*time.Hour {
-		return fmt.Sprintf("%.0f hours", d.Hours())
+		return fmt.Sprintf("%d hours", int(d.Hours()))
 	}
-	return fmt.Sprintf("%.0f days", d.Hours()/24)
+	return fmt.Sprintf("%d days", int(d.Hours()/24))
 }
 
 // deadFindings 找出累计命中为零的规则。
